Split the canvas into the concerns it turned out to have
PreviewCanvas had grown to 756 lines holding eight jobs at once, which is
the shape a feature takes when it is found by debugging rather than
designed. Each job now has a name and a file, and the component is 534
lines of which about a third is the render:

- useCanvasViewport — how big the visible box is, and why measuring the
  wrong element runs the surface off to millions of pixels.
- useCanvasFrames — the frames as live documents: the registry, what each
  has been told, how long its page is, what it hands back.
- useCanvasPlacement — where the canvas sits and who decided that, with
  the three rules that make it impossible to lose.

No behaviour change; the 44 canvas tests and the whole suite pass
unchanged.

PAGE HEIGHT. Growing monotonically was wrong in the other direction: a
transient measured mid-load stuck, leaving a frame with a screen of white
space under a page that was actually shorter. A height must now be
measured twice in a row before it counts — an oscillating pair never
repeats, so it still cannot flicker — and it is measured from where the
content ends rather than from `scrollHeight`, which reports the frame's
own height back once the body is stretched to it.
